+++ src/main/java/gift/controller/view/LoginViewController.java
@@ -1,4 +1,4 @@
-package gift.yjshop.controller;
+package gift.controller.view;
 
 import gift.dto.MemberRequestDto;
 import gift.dto.Role;
@@ -47,7 +47,6 @@ public class LoginViewController {
             HttpServletResponse response,
             HttpServletRequest request
     ){
-
         if(memberService.getMemberByEmail(memberRequestDto.email()).isPresent()){
             bindingResult.addError(new FieldError("memberRequestDto", "email", "이미 사용중인 이메일 입니다."));
         }
@@ -56,12 +55,7 @@ public class LoginViewController {
             response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
             return "/yjshop/user/register";
         }
-
-        Member member = memberService.register(memberRequestDto);
-        request.setAttribute("memberId", member.getMemberId());
-        request.setAttribute("role", member.getRole());
-        response.setStatus(HttpServletResponse.SC_CREATED);
-
+        memberService.register(memberRequestDto);
         return "redirect:/view/loginform";
     }
 
@@ -76,20 +70,16 @@ public class LoginViewController {
     //로그인 기능 -> 토큰을 반환(쿠키에 저장)
     @PostMapping("/login")
     public String login(HttpServletRequest request) {
-
         Role role = (Role) request.getAttribute("role");
-
         //관리자 로그인 시, -> 관리자 페이지로 이동
         if(role.equals(Role.ADMIN)){
             return "redirect:/view/admin/products";
         }
         return "redirect:/view/products/list";
-
     }
 
     @PostMapping("/login/error")
     public String loginError(HttpServletRequest request, Model model){
-        System.out.println("request = " + request.getAttribute("errormsg"));
         model.addAttribute("errormsg", request.getAttribute("errormsg"));
         return "/yjshop/user/loginerror";
     }
