+++ src/main/java/gift/infra/config/WebConfig.java
@@ -28,9 +28,9 @@ public class WebConfig implements WebMvcConfigurer {
 
     public void addInterceptors(InterceptorRegistry registry) {
         registry.addInterceptor(loginCheckInterceptor)
-                .addPathPatterns("/api/wishlist/**");
+                .addPathPatterns("/view/my/**");
         registry.addInterceptor(adminCheckInterceptor)
-                .addPathPatterns("/api/admin/**");
+                .addPathPatterns("/view/admin/**");
     }
 
     @Override
