+++ src/main/java/gift/infra/filter/FilterExceptionHandler.java
@@ -3,6 +3,7 @@ package gift.infra.filter;
 
 import gift.exception.MyException;
 import jakarta.servlet.FilterChain;
+import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -19,11 +20,10 @@ public class FilterExceptionHandler extends OncePerRequestFilter {
         try {
             filterChain.doFilter(request, response);
         } catch (MyException e) {
-            logger.info("[FilterExceptionHandler]"+ e.getErrorCode());
-            String msg = e.getErrorCode().getMessage();
-            response.setStatus(e.getErrorCode().getStatusCode().value());
-            response.setContentType("application/json;charset=UTF-8");
-            response.getWriter().write("{\"message\":\"" + msg + "\"}");
+            logger.info("[FilterExceptionHandlerForView]" + e.getErrorCode());
+            request.setAttribute("errormsg", e.getErrorCode().getMessage());
+            RequestDispatcher dispatcher = request.getRequestDispatcher("/view/login/error");
+            dispatcher.forward(request, response);
         }
     }
 }
