+++ src/main/java/gift/infra/filter/LoginFilter.java
@@ -14,6 +14,7 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletInputStream;
 import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.BufferedReader;
@@ -28,12 +29,10 @@ public class LoginFilter implements Filter {
     private static final Logger log = LoggerFactory.getLogger(LoginFilter.class);
     private final MemberService memberService;
     private final JwtAuthService jwtAuthService;
-    private final ObjectMapper objectMapper;
 
     public LoginFilter(MemberService memberService, JwtAuthService jwtAuthService, ObjectMapper objectMapper){
         this.memberService = memberService;
         this.jwtAuthService = jwtAuthService;
-        this.objectMapper = objectMapper;
     }
 
     @Override
@@ -41,6 +40,7 @@ public class LoginFilter implements Filter {
         log.info("필터 등록 완료,,,");
     }
 
+    //로그인 요청에 대해서만 동작하는 필터임...
     @Override
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
             throws IOException, ServletException {
@@ -51,20 +51,16 @@ public class LoginFilter implements Filter {
         String url = httpServletRequest.getRequestURI();
         String Method = httpServletRequest.getMethod();
 
-        //로그인 요청에 대해서만 동작하는 필터임...
-        if(url.equals("/api/members/login") && Method.equals("POST")){
+        //view로 로그인 요청이 들어오는 경우
+        if(url.equals("/view/login") && Method.equals("POST")){
 
             log.info("로그인 필터 is working,,,,");
 
-            BufferedReader inputStream = httpServletRequest.getReader();
-            MemberRequestDto memberRequestDto = objectMapper.readValue(inputStream, MemberRequestDto.class);
-
-            //HttpServletRequest에서 인증정보를 파싱하고
-            String email = memberRequestDto.email();
-            String password = memberRequestDto.password();
+            String email = request.getParameter("email");
+            String password = request.getParameter("password");
 
             //인증실패예외를 반환하고 이를 Http Response로 렌더링하는 작업이 필요할것 같아요.
-            if(email == null || password == null){
+            if(email.isBlank()|| password.isBlank()){
                 throw new MyException(ErrorCode.EMAIL_PASSWORD_REQUIRED);
             }
 
@@ -76,9 +72,16 @@ public class LoginFilter implements Filter {
             Member member = memberService.getMemberByEmail(email).get();
             String token = jwtAuthService.createJwt(email, member.getMemberId(), member.getRole());
 
-            httpServletResponse.addHeader("Authorization", token);
+            //쿠키 발행 (쿠키에 토큰을 저장)
+            Cookie tcookie = new Cookie("token", token);
+            tcookie.setPath("/");
+            httpServletResponse.addCookie(tcookie);
+
+            request.setAttribute("token", token);
+            request.setAttribute("role", member.getRole());
             log.info("토큰 생성 완료");
         }
+
         //다음 필터가 있다면 동작해라;
         chain.doFilter(request, response);
     }
