+++ src/main/java/gift/infra/interceptor/LoginCheckInterceptor.java
@@ -3,6 +3,7 @@ package gift.infra.interceptor;
 import gift.exception.ErrorCode;
 import gift.exception.MyException;
 import gift.service.JwtAuthService;
+import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
@@ -23,21 +24,35 @@ public class LoginCheckInterceptor implements HandlerInterceptor {
 
     //컨트롤러에 호출 전에 동작
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        log.info("로그인 정보를 확인 중,,,");
-        log.info("LoginCheckInterceptor.preHandle");
+        //컨트롤러가 호출되기 전에 실행됨
+        log.info("[LoginChecker] preHandle");
+        String token = null;
 
-        String token = request.getHeader("Authorization"); //헤더에서 토큰을 꺼내서
-
-        //토큰이 없는 경우
-        if(token == null){
-            //로그인하는 페이지로 유도
-            throw new MyException(ErrorCode.LOGIN_REQUIRED_FAIL);
+        try{
+            Cookie[] cookies = request.getCookies();
+            for(Cookie c : cookies){
+                if(c.getName().equals("token")){
+                    token = c.getValue();
+                }
+            }
+        }
+        catch (NullPointerException e){
+            log.info("쿠키가 존재하지 않음"); //쿠키가 존재하지 않는 경우,,,
+            response.sendRedirect("/view/loginform"); //로그인 화면으로 유도
+            return false; //컨트롤러가 동작하지 않는다.
         }
 
-        //토큰이 존재하는 경우 -> 토큰을 검증
+        //쿠키에 토큰이 없다면 -> 로그인 페이지로 이동시키기
+        if(token == null){
+            response.sendRedirect("/view/loginform");
+            return false; //컨트롤러가 동작하지 않는다.
+        }
+
+        //토큰의 유효성을 확인하기
+        log.info("로그인 정보를 확인 중,,,");
         jwtAuthService.checkValidation(token);
         log.info("JWT 토큰 검증 성공,,,");
-        return true;
+        return true; //컨트롤러가 동작
     }
 
 }
