+++ src/main/java/gift/service/JwtAuthService.java
@@ -11,7 +11,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
 //JWT와 관련된 서비스 :
-@Slf4j
 @Service
 public class JwtAuthService {
 
@@ -19,8 +18,7 @@ public class JwtAuthService {
     private final String secretKey = "ComeOnYouGunnersNorthLondonisRedNorthLondonFOREVER";
 
     //payload의 정보를 추출하는 함수
-    public Long getMemberId(String bearerToken){
-        String token = bearerToken.split(" ")[1]; //접두사 제거
+    public Long getMemberId(String token){
         return Jwts.parser()
                 .verifyWith(Keys.hmacShaKeyFor(secretKey.getBytes()))
                 .build()
@@ -30,8 +28,7 @@ public class JwtAuthService {
     }
 
     //payload의 정보를 추출하는 함수
-    public String getMemberRole(String bearerToken){
-        String token = bearerToken.split(" ")[1]; //접두사 제거
+    public String getMemberRole(String token){
         return Jwts.parser()
                 .verifyWith(Keys.hmacShaKeyFor(secretKey.getBytes()))
                 .build()
@@ -40,7 +37,7 @@ public class JwtAuthService {
                 .get("role", String.class);
     }
 
-    //TODO: 토큰 생성
+    //토큰 생성
     public String createJwt(String email, Long memberId, Role role){
         return Jwts.builder()
                 .claim("email", email)
@@ -50,15 +47,12 @@ public class JwtAuthService {
                 .compact();
     }
 
-    //TODO: 토큰을 검증 -> 로그인 이후의 동작 (wishList -> 사용자별 wishList 존재)
-    public void checkValidation(String bearerToken) {
-
-        if(bearerToken==null || !bearerToken.startsWith("Bearer ")){
+    //토큰을 검증 -> 로그인 이후의 동작 (wishList -> 사용자별 wishList 존재)
+    public void checkValidation(String token) {
+        if(token==null){
             throw new MyException(ErrorCode.JWT_VALIDATION_FAIL);
         }
-
         try{
-            String token = bearerToken.split(" ")[1]; //Bearer 접두사 제거
             Jwts.parser().verifyWith(Keys.hmacShaKeyFor(secretKey.getBytes())).build().parseSignedClaims(token);
         }
         catch (Exception e){
