--- src/main/resources/templates/members/modifyForm.html
@@ -1,26 +1,0 @@
-<html layout:decorate="~{members/layout}">
-<div layout:fragment="content">
-  <div id="productForm">
-    <form th:action="@{/admin/members/modify/{id}(id=${member.memberId})}" th:method="post" th:object="${memberRequestDto}">
-      <p>
-        <h3>상품 수정하기</h3>
-      </p>
-      <p style="display: flex; align-items: center; gap: 10px;">
-        <strong>이메일</strong>
-        <input type="email" name="email" placeholder="이메일 입력" th:value="${member.email}">
-        <span th:if="${#fields.hasErrors('email')}" th:errors="*{email}" style="color: crimson;">Email Error</span>
-      </p>
-      <p>
-        <strong>비밀번호</strong>
-        <input type="text" name="password" placeholder="비밀번호 입력" th:value="${member.password}">
-        <span th:if="${#fields.hasErrors('password')}" th:errors="*{password}" style="color: crimson;">Password Error</span>
-      </p>
-      <p>
-        <input type="submit" value="회원 정보 수정하기">
-        <button type="button" th:onclick="|location.href='@{/admin/members/list}'|">수정 취소</button></td>
-      </p>
-    </form>
-  </div>
-
-</div>
-</html>
