--- src/main/resources/templates/productinfo.html
@@ -1,33 +1,0 @@
-<html layout:decorate="~{layout}">
-<div layout:fragment="content">
-
-  <table>
-    <thead>
-    <tr>
-      <th>상품ID</th>
-      <th>상품명</th>
-      <th>가격</th>
-      <th>이미지</th>
-      <th>Actions</th>
-    </tr>
-    </thead>
-
-    <tbody>
-    <tr>
-      <td th:text="${product.id}"></td>
-      <td th:text="${product.name}"></td>
-      <td th:text="${product.price}"></td>
-      <td><img th:src="${product.imageUrl}" width="100px" alt = "상품 이미지"></td>
-      <td>
-        <form th:action="@{/admin/products/remove/{id}(id=${product.id})}" method="post">
-          <button type="submit">상품 삭제</button>
-        </form>
-        <button th:onclick="|location.href='@{/admin/products/modify/{id}(id=${product.id})}'|">상품 수정</button>
-      </td>
-
-    </tr>
-    </tbody>
-  </table>
-
-</div>
-</html>
